requirements.txt
langchain==0.3.27
langchain_openai==0.3.28
langchain_anthropic==0.3.18
langchain_google_genai==2.1.9
python-dotenv==1.1.1
